requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0